client/requirements.txt
Flask==3.0.3
httpx==0.27.0
langchain_community==0.2.6
langchain_core==0.2.10
langchain_openai==0.1.13
langgraph==0.1.5
loguru==0.7.2
openai==1.35.7
pandas==2.2.2
# protobuf==5.27.2
protobuf==4.25.3
PyPDF2==3.0.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
PyYAML==6.0.1
streamlit==1.35.0
streamlit_authenticator==0.3.2
streamlit_option_menu==0.3.13
tavily_python==0.3.3
twilio==9.2.1
google-cloud-firestore==2.15.0